scraper/requirements.txt
# Python dependencies for property scraper
scrapy>=2.13.3
psycopg2-binary>=2.9.6
requests>=2.31.0
PyYAML>=6.0
python-dotenv>=1.0.0
# Browser automation for JavaScript sites
selenium>=4.15.0
scrapy-selenium>=0.4.2
